refactor(PointResource): 오목판의 특정 좌표의 ImageResource를 반환하는 책임을 MainActivity.kt 파일에서 PointResource.kt 파일로 옮김
